pyproject.toml
[tool.poetry]
name = "shopping-cart"
version = "0.1.0"
description = ""
authors = ["Venu Vardhan Reddy Tekula <venuvardhanreddytekula8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.5"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3.2"
flake8 = "^4.0.1"
behave-django = "^1.4.0"
selenium = "^4.1.3"
Pattern = "^3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
makemigrations = "scripts:makemigrations"
migrate = "scripts:migrate"
server = "scripts:server"
shell = "scripts:shell"
tests = "scripts:tests"
createsuperuser = "scripts:createsuperuser"
